# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Estatísticas","description":"Consulta de estatísticas calculadas sobre as transações registradas"},{"name":"Transações","description":"Operações relacionadas ao gerenciamento de transações"}],"paths":{"/transacao":{"post":{"tags":["Transações"],"summary":"Adiciona uma nova transação","description":"Registra uma transação financeira na memória. O valor não pode ser negativo e a data não pode ser no futuro.","operationId":"adicionar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransacaoRequest"}}},"required":true},"responses":{"201":{"description":"Transação criada com sucesso","content":{"application/json":{"schema":{"type":"string"}}}},"422":{"description":"Dados inválidos: valor negativo ou data/hora no futuro","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Erro genérico na requisição, como JSON malformado","content":{"application/json":{"schema":{"type":"string"}}}}}},"delete":{"tags":["Transações"],"summary":"Remove todas as transações","description":"Limpa todas as transações armazenadas em memória. A operação não pode ser desfeita.","operationId":"limpar","responses":{"200":{"description":"Transações removidas com sucesso","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/estatistica":{"get":{"tags":["Estatísticas"],"summary":"Retorna estatísticas das transações","description":"Calcula e retorna count, média, mínimo e máximo dos valores das transações registradas dentro da janela de tempo informada.","operationId":"estatistica","parameters":[{"name":"intervaloEmSegundos","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":60}}],"responses":{"200":{"description":"Estatísticas calculadas com sucesso","content":{"application/json":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"TransacaoRequest":{"type":"object","properties":{"valor":{"type":"number","description":"Valor da transação em reais. Não pode ser negativo.","example":150.75},"dataHora":{"type":"string","format":"date-time","description":"Data e hora em que a transação ocorreu, com fuso horário. Não pode ser no futuro.","example":"2024-03-15T10:30:00-03:00"}}}}}}